refactor: split the crate into a lib and bin shims
