+++ src/bin/spotatui.rs
@@ -1,0 +1,6 @@
+use anyhow::Result;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+  spotatui::run_cli().await
+}
